SudokuBoard Cloning, Locked Status for Coordinates, and Beginnings of Solver
Implemented a way to clone a sudoku board.  Relies on text entry of board.  Feels like a clever soluation atm, but it may result in future problems.  Consider finding a different way.

Implemented a locked status for coordinates.  Locked status allows me to identify which numbers were given in the input and which ones the solver has changed.

Began putting the solver together.  Next step is to write methods to check the x, y, and square sections.